bookmark/pyproject.toml
[tool.poetry]
name = "bookmark"
version = "0.1.0"
description = ""
authors = ["PJG8806 <jinguy39@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
ipython = "^9.9.0"
django-extensions = "^4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
